Updated the number of options.
IPL Score Prediction 
In today’s date data analysis is need for every data analytics to examine the sets of data to extract the
useful information from it and to draw conclusion according to the information. Data analytics
techniques and algorithms are more used by the commercial industries which enables them to take
precise business decisions. It is also used by the analysts and the experts to authenticate or negate
experimental layouts, assumptions and conclusions. In recent years the analytics is being used in the
field of sports to predict and draw various insights. Due to the involvement of money, team spirit,
city loyalty and a massive fan following, the outcome of matches is very important for all stake
holders. In this paper, the past seven year’s data of IPL containing the player’s details, match venue
details, teams, ball to ball details, is taken and analyzed to draw various conclusions which help in
the improvement of a player’s performance. Various other features like how the venue or toss decision
has influenced the winning of the match in last seven years are also predicted. Various machine
learning and data extraction models are considered for prediction are Linear regression, Decision tree,
K-means, Logistic Regression etc. The cross validation score and the accuracy are also calculated
using various machine learning algorithms. Before prediction we have to explore and visualize the
data because data exploration and visualization is an important stage of predictive modeling.